init: Set-up of the new structure with minimal coverage for grace.h
New branch to set up the skeleton of what could be the new structure of the project. This structure follows the standard of having a lib folder and the associated tests. In this case:
* Lib folder = dft_lib
* Tests folder = test
We are using the following standards:
* Folder structure: https://github.com/google/googletest/tree/master/googletest
* Coding style: https://google.github.io/styleguide/cppguide.html
The testing will be done by using Google's library GTests (https://github.com/google/googletest/tree/master/googletest).