#5 Add Close method
Fix: Minor mistake in the conditions to throw exception in the setters of XLimits and YLimits: limits can be negative indeed

Fix: Use of Close method in destructor, which checks whether xmgrace was initialised. This removes an error message due to the fact that there is no opened xmgrace

Feat: Reorganised tests in regions according to the namespaces. Add new unit tests for the new setters and close method